Improve removing node with two children
Previously, if the node had two children we'd remove one of the children and copy its value into the node you called remove() on. The advantage of that approach is that calling remove() on the root of the tree doesn't invalidate the root node. However, this wasn't true for when you tried to delete a root node that has only one child. That was confusing.

Now we always truly remove the node you called remove() on, and tell you what its replacement is. As a consequence, if you remove the root node you must make sure to use its replacement node as the new root from then on.