Refactors Binary Search implementation to be more Swifty.
This involved converting the global function into methods of
`Collection`. By doing so, the recursive version no longer needs to
take a search range as a parameter. It also made both the recursive and
iterative versions of the algorithm usable on other collection types
besides `Array`. Tests were also refactored to test both versions of
the algorithm.