fix delete
By having a BST that is defined by using recursive BSTs you limit the ability to have an empty tree.
The current code has an issue not only with the empty tree, but also with deleting the root node.  Currently, if you delete the root node, the entire tree is deleted except for the root.
By using a separate BSTNode class, the tree itself can have a reference to the root node and do all operations from there.  Some of the variables and methods got moved to the BSTNode, although it may be better to reference some of the methods from the tree using the root.
The parent node instance of the BSTNode isn't needed for delete.  It was left in for some of the other methods that use it.
The corresponding other playgrounds that reference the BST along with the tutorial will need to be updated as well.  Please let me know if you would like me to work on that as part of this PR.
Cheers, Eric